We may need to run autoreconf twice in bootstrap
The first time you run autoreconf (which runs automake), automake will
not find config/depconf (because it doesn't exist yet), and will not
put it into DIST_COMMON in Makefile.in.  So if it's not in there, run
autoreconf again.

With this fix, "make dist" is reproducible.  Untarring the tarball it
produces, and running "./configure; make dist" then produces a
byte-for-byte identical tarball.